requirements.txt
psycopg2-binary
sentence-transformers
faker
python-dotenv
numpy
